Securing some of the Views

+++ products/views.py
@@ -1,5 +1,7 @@
+from django.db.models.functions import Lower
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
 from products.forms import ProductForm
@@ -66,8 +68,13 @@ def find_product(request, sku):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -86,8 +93,14 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, sku):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, sku=sku)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -110,8 +123,13 @@ def edit_product(request, sku):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, sku):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, sku=sku)
     product.delete()
     messages.success(request, 'Product deleted!')
